refactor: extract safe ops into dedicated microcrate
