+++ copybook-safe-ops/README.md
@@ -1,0 +1,6 @@
+# copybook-safe-ops
+
+Panic-safe primitive operations extracted from `copybook-utils`.
+
+This crate provides reusable checked parsing, arithmetic, string formatting, and bounds-safe helpers
+for high-throughput copybook processing.
